The basic int64 change for the B+Tree has been implemented and all core B+Tree tests are passing, as are the WORM tests and the SAIL quads test suite.  I am going to commit and see how this looks under CI.
git-svn-id: svn+ssh://svn.code.sf.net/p/bigdata/code/branches/INT64_BRANCH@4494 8f7bc0f5-282e-0410-95e3-8d296e9bb460

+++ bigdata-rdf/src/test/com/bigdata/rdf/spo/TestXXXCShardSplitHandler.java
@@ -310,11 +310,11 @@ public class TestXXXCShardSplitHandler extends AbstractTestSegSplitter {
                  */
                 
                 // Get insertion point (there are no 8-byte keys in the ndx).
-                final int pos = seg.indexOf(separatorKey);
+                final long pos = seg.indexOf(separatorKey);
                 assertTrue(pos < 0);
                 
                 // Convert to a tuple index.
-                final int index = -(pos) - 1;
+                final long index = -(pos) - 1;
 
                 if (log.isInfoEnabled())
                     log.info("index=" + index);
